feat: HashedByteArray -> BinaryNBTCompound
